Add the QuestBitDetails screen when clicking "View Details" Removing the Refresh button (make the QuestBits update automatically) -- maybe adding a short loading screen will help prevent the bug
