--- QuestBit/android/app/components/QuestBitDetails.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import { View } from 'react-native';
-import { Text } from 'react-native-elements';
-import { RouteProp } from '@react-navigation/native';
-import { RootStackParamList } from '../../../App';
-
-const QuestBitDetails = () => {
-
-    return (
-        <View>
-           <Text>In View Details page!</Text>
-        </View>
-    );
-}
-
-export default QuestBitDetails;
